more ai changes
