+++ classes/class-ai-tasks.php
@@ -77,60 +77,6 @@ class AI_Tasks {
 	}
 
 	/**
-	 * Execute an AI task.
-	 *
-	 * @param int    $task_id     The task ID to execute.
-	 * @param string $site_url    The site URL to analyze.
-	 * @param string $license_key The license key.
-	 * @return array|WP_Error Response array on success, WP_Error on failure.
-	 */
-	public function execute_ai_task( $task_id, $site_url, $license_key ) {
-		$url = \progress_planner()->get_remote_server_root_url() . '/wp-json/progress-planner-saas/v1/execute-ai-task';
-
-		// Testing.
-		$site_url = 'https://progressplanner.com';
-
-		$post_data = [
-			'task_id'     => $task_id,
-			'site_url'    => $site_url,
-			'license_key' => $license_key,
-		];
-
-		$response = \wp_remote_post(
-			$url,
-			[
-				'body'    => $post_data,
-				'timeout' => 45, // AI tasks may take longer.
-			]
-		);
-
-		if ( \is_wp_error( $response ) ) {
-			\error_log( 'WP_Error response: ' . $response->get_error_message() );
-			return $response;
-		}
-
-		$response_code = \wp_remote_retrieve_response_code( $response );
-		$body          = \wp_remote_retrieve_body( $response );
-		$json          = \json_decode( $body, true );
-
-		// Debug logging for response.
-		\error_log( 'SaaS server response:' );
-		\error_log( '  Response code: ' . $response_code );
-		\error_log( '  Response body: ' . $body );
-
-		if ( 200 !== $response_code ) {
-			$error_message = isset( $json['message'] ) ? $json['message'] : 'Unknown error occurred';
-			return new \WP_Error( 'ai_execution_failed', $error_message, [ 'status' => $response_code ] );
-		}
-
-		if ( ! \is_array( $json ) ) {
-			return new \WP_Error( 'invalid_response', 'Invalid response from server', [ 'status' => 500 ] );
-		}
-
-		return $json;
-	}
-
-	/**
 	 * Get a cached AI response for a task.
 	 *
 	 * @param int $task_id The task ID.
@@ -138,7 +84,7 @@ class AI_Tasks {
 	 */
 	public function get_cached_response( $task_id ) {
 		$cache_key = 'prpl_ai_response_' . $task_id;
-		return progress_planner()->get_utils__cache()->get( $cache_key );
+		return \progress_planner()->get_utils__cache()->get( $cache_key );
 	}
 
 	/**
@@ -151,7 +97,7 @@ class AI_Tasks {
 	 */
 	public function cache_response( $task_id, $response, $expiry = WEEK_IN_SECONDS ) {
 		$cache_key = 'prpl_ai_response_' . $task_id;
-		return progress_planner()->get_utils__cache()->set( $cache_key, $response, $expiry );
+		return \progress_planner()->get_utils__cache()->set( $cache_key, $response, $expiry );
 	}
 
 	/**
@@ -162,6 +108,6 @@ class AI_Tasks {
 	 */
 	public function clear_cached_response( $task_id ) {
 		$cache_key = 'prpl_ai_response_' . $task_id;
-		return progress_planner()->get_utils__cache()->delete( $cache_key );
+		return \progress_planner()->get_utils__cache()->delete( $cache_key );
 	}
 }
